pyproject.toml
[project]
name = "mcp-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "python-dotenv>=1.2.2",
]
